requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
pymysql
python-dotenv
cryptography
